add new use case to check for app update and implement it
