+++ app/src/main/java/org/groundplatform/android/ui/main/MainActivity.kt
@@ -161,7 +161,7 @@ class MainActivity : AbstractActivity() {
 
   override fun onResume() {
     super.onResume()
-    if (viewModel.isAppUpdateAvailable()) {
+    if (viewModel.isAppUpdateRequired()) {
       showForceUpdateDialog()
     }
   }
